pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-infra-radar"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.34,<1.0",
  "typer>=0.12,<1.0",
  "pydantic>=2.9,<3.0",
  "PyYAML>=6.0,<7.0",
  "SQLAlchemy>=2.0,<3.0",
  "httpx>=0.28,<1.0",
  "beautifulsoup4>=4.12,<5.0",
  "apscheduler>=3.10,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<10.0",
  "respx>=0.20,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["radar*"]
